refactor: Push trim to string utilities
